[hotfix][table-planner] Fix integral types list in FunctionGenerator

+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/codegen/calls/FunctionGenerator.scala
@@ -33,7 +33,7 @@ import scala.collection.mutable
 object FunctionGenerator {
 
   val INTEGRAL_TYPES = Array(
-    INTEGER,
+    TINYINT,
     SMALLINT,
     INTEGER,
     BIGINT)
